Specs for erlang config.
The specs work OK when run on their own, but rspec-puppet/issues/44
makes them fail when run together...

Also, templating may fail or not fail, depending on the same
call to loadall in the autoloader, so I put a begin ; ... ; rescue ; end
block in there, to shut it up. Of course this isn't the best way forward,
and if the underlying cause is fixed, then I'm happy to remove that.