Allow ulimits to be managed externally
I like to manage my /etc/security/limits.conf file for my servers (as
it's not specific to riak) via another method.

This commit makes it possible to disable the riak module's default
behaviour of trying to manage the /etc/security/limits.conf file, which
will cause a resource conflict if also defined in another module.