Allow installing curl to be optional
curl is a very common package to install. If this module manages curl
and another module also attempts to manage the package['curl'] resource,
then there will be a resource conflict.

This allows optionally disabling the management of the curl package,
should the user wish to manage it via another module.